dosc: SinpupDetailPage, MapCompnent로 분리
