+++ src/Pages/SignpDetailPage/SingupDetailPage.tsx
@@ -2,7 +2,8 @@ import axios from 'axios';
 import { useState, useCallback, useEffect } from 'react'
 import { useNavigate } from 'react-router-dom';
 import { REST_API_KEY, REDIRECT_URI } from '../LoginPage/kakao/kakaoLoginData';
-import { Container, InputBox, Title, Box, EmailInputBox, EmailInput, EmailCheckBtn, Alertment, PhonenumberInput, MapBox, Map, UserLastBtn} from './styled'
+import { Container, InputBox, Title, Box, EmailInputBox, EmailInput, EmailCheckBtn, Alertment, PhonenumberInput, MapBox, UserLastBtn} from './styled'
+import MapComponent from './MapComponent';
 
 declare global {
   interface Window {
@@ -51,70 +52,16 @@ const SingupDetailPage: React.FC = () => {
     }
   }, [])
 
-  useEffect(() => {
-    const mapContainer = document.getElementById('map'); // 지도를 표시할 div
-    const mapOption = {
-      center: new window.kakao.maps.LatLng(33.450701, 126.570667), // 지도의 중심좌표
-      level: 5, // 지도의 확대 레벨
-    };
-
-    const map = new window.kakao.maps.Map(mapContainer, mapOption); // 지도를 생성합니다
-
-    // HTML5의 geolocation으로 사용할 수 있는지 확인합니다
-    if (navigator.geolocation) {
-      // GeoLocation을 이용해서 접속 위치를 얻어옵니다
-      navigator.geolocation.getCurrentPosition(function (position) {
-        const lat = position.coords.latitude; // 위도
-        const lon = position.coords.longitude; // 경도
-
-        const locPosition = new window.kakao.maps.LatLng(lat, lon); // 마커가 표시될 위치를 geolocation으로 얻어온 좌표로 생성합니다
-        const message = '<div style="padding:5px;">여기에 계신가요?!</div>'; // 인포윈도우에 표시될 내용입니다
-
-        // 마커와 인포윈도우를 표시합니다
-        displayMarker(locPosition, message);
-      });
-    } else {
-      // HTML5의 GeoLocation을 사용할 수 없을 때 마커 표시 위치와 인포윈도우 내용을 설정합니다
-      const locPosition = new window.kakao.maps.LatLng(33.450701, 126.570667);
-      const message = 'geolocation을 사용할 수 없어요..';
-
-      // 마커와 인포윈도우를 표시합니다
-      displayMarker(locPosition, message);
-    }
-
-    // 지도에 마커와 인포윈도우를 표시하는 함수입니다
-    function displayMarker(locPosition: any, message: string) {
-      // 마커를 생성합니다
-      const marker = new window.kakao.maps.Marker({
-        map: map,
-        position: locPosition,
-      });
-
-      const iwContent = message; // 인포윈도우에 표시할 내용
-      const iwRemoveable = true;
-
-      // 인포윈도우를 생성합니다
-      const infowindow = new window.kakao.maps.InfoWindow({
-        content: iwContent,
-        removable: iwRemoveable,
-      });
-
-      // 인포윈도우를 마커위에 표시합니다
-      infowindow.open(map, marker);
-
-      // 지도 중심좌표를 접속위치로 변경합니다
-      map.setCenter(locPosition);
-    }
-  }, []);
-
-
   // 카카오 로그인 사용자의 정보를 얻음
   const [kakaoToken, setKakaoToken] = useState<any>(null)
   const [kakaoUserDoc, setKakaoUserDoc] = useState<any>(null)
 
   const params = new URL(document.location.toString()).searchParams
+
+  // code => 카카오 로그인 과정에서 인증 후에 발급된 인증 코드
   const code = params.get('code')
 
+  // 전달받은 인증 코드를 사용하여 카카오 서버로부터 액세스 토큰을 요청함
   const getToken = async (code: string) => {
     const grant_type = 'authorization_code'
     const client_id = `${REST_API_KEY}`
@@ -129,7 +76,9 @@ const SingupDetailPage: React.FC = () => {
     )
 
     const token = res.data.access_token
-    setKakaoToken(token)
+    // token => 엑세스 토큰 
+    setKakaoToken(token) 
+    // kakaoToken => 액세스 토큰
   }
 
   const getKaKaoUserData = async (kakaoToken: any) => {
@@ -178,7 +127,6 @@ const SingupDetailPage: React.FC = () => {
   }
 
 
-
   return (
     <Container>
       <InputBox>
@@ -204,14 +152,12 @@ const SingupDetailPage: React.FC = () => {
         </Box>
 
         <MapBox>
-          <Map id="map" style={{ width: '80%', height: '250px' }} />
+          <MapComponent mapCenter={{ lat: 33.450701, lon: 126.570667 }} />
         </MapBox>
         
         {/* 닉네임, 전화번호, 위치 User DB 에 저장 
           메인페이지로 가는 임시 버튼 */}
         <UserLastBtn onClick={setingUser}>현재 위치로 등록</UserLastBtn>
-
-
       </InputBox>
     </Container>
   )
